chore: rename compare-files
