+++ test/browserify.test.js
@@ -13,7 +13,7 @@ var fs     = require("fs"),
     plugin = require("../src/browserify"),
     
     bundle  = require("./lib/bundle"),
-    compare = require("./lib/compare-files");
+    compare = require("./lib/compare.js");
 
 describe("/browserify.js", function() {
     describe("basic functionality", function() {
