+++ test/lib/compare.js
