+++ test/processor.methods.test.js
@@ -6,7 +6,7 @@ var fs      = require("fs"),
     
     Processor = require("../src/processor"),
     
-    compare = require("./lib/compare-files");
+    compare = require("./lib/compare.js");
 
 describe("/processor.js", function() {
     describe("Methods", function() {
